LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/iframe_sandbox_anchor_download_block_downloads.tentative-expected.txt

CONSOLE MESSAGE: Not allowed to download due to sandboxing
CONSOLE MESSAGE: Not allowed to download due to sandboxing
CONSOLE MESSAGE: Not allowed to download due to sandboxing


PASS <a download> triggered download in sandbox is blocked.
PASS <a download> triggered download in sandbox is blocked before a request is made.
FAIL <a > triggered download in sandbox is blocked. assert_equals: Expect no download to happen, but got one. expected "TOKEN_NOT_SET" but got "TOKEN_SET"
FAIL <a target="_blank" > triggered download in sandbox is blocked. assert_equals: Expect no download to happen, but got one. expected "TOKEN_NOT_SET" but got "TOKEN_SET"
FAIL <a target="_blank" rel="noopener" > triggered download in sandbox is blocked. assert_equals: Expect no download to happen, but got one. expected "TOKEN_NOT_SET" but got "TOKEN_SET"
